prolaze testovi portrait ui gotov

+++ app/src/main/java/ba/etf/rma23/projekat/data/repositories/SavedGame.kt
@@ -3,6 +3,6 @@ package ba.etf.rma23.projekat.data.repositories
 import com.google.gson.annotations.SerializedName
 
 data class SavedGame(
-    @SerializedName("igdb_id") val id : Long,
+    @SerializedName("igdb_id") val id : Int,
     @SerializedName("name") val title : String
 )
